chore: update CredentialProvider abstract class and StringMomentoTokenProvider class
