+++ src/Auth/EnvMomentoTokenProvider.php
@@ -4,11 +4,12 @@ declare(strict_types=1);
 namespace Momento\Auth;
 
 use Momento\Cache\Errors\InvalidArgumentError;
+use function Momento\Utilities\isNullOrEmpty;
 
 /**
  * Reads and parses a JWT token stored as an environment variable.
  */
-class EnvMomentoTokenProvider extends CredentialProvider
+class EnvMomentoTokenProvider extends StringMomentoTokenProvider
 {
     public function __construct(
         string  $envVariableName,
@@ -18,7 +19,11 @@ class EnvMomentoTokenProvider extends CredentialProvider
         ?string $trustedCacheEndpointCertificateName = null
     )
     {
-        parent::__construct($envVariableName, $controlEndpoint, $cacheEndpoint, $trustedControlEndpointCertificateName, $trustedCacheEndpointCertificateName);
+        $authToken = getenv($envVariableName);
+        if ($authToken === false || isNullOrEmpty($authToken)) {
+            throw new InvalidArgumentError("Environment variable $envVariableName is empty or null.");
+        }
+        parent::__construct($authToken, $controlEndpoint, $cacheEndpoint, $trustedControlEndpointCertificateName, $trustedCacheEndpointCertificateName);
     }
 
     /**
