+++ src/Auth/StringMomentoTokenProvider.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace Momento\Auth;
 
 use Momento\Cache\Errors\InvalidArgumentError;
+use function Momento\Utilities\isNullOrEmpty;
 
 /**
  * Reads and parses a JWT token stored as a string.
@@ -18,7 +19,46 @@ class StringMomentoTokenProvider extends CredentialProvider
         ?string $trustedCacheEndpointCertificateName = null
     )
     {
-        parent::__construct($authToken, $controlEndpoint, $cacheEndpoint, $trustedControlEndpointCertificateName, $trustedCacheEndpointCertificateName);
+        if (isNullOrEmpty($authToken)) {
+            throw new InvalidArgumentError("String $authToken is empty or null.");
+        }
+        if ($trustedControlEndpointCertificateName xor $trustedCacheEndpointCertificateName) {
+            throw new InvalidArgumentError(
+                "If either of trustedCacheEndpointCertificateName or trustedControlEndpointCertificateName " .
+                "are provided, they must both be."
+            );
+        }
+        $this->authToken = $authToken;
+        $payload = AuthUtils::parseAuthToken($authToken);
+        $this->controlEndpoint = $controlEndpoint ?? $payload->cp;
+        $this->cacheEndpoint = $cacheEndpoint ?? $payload->c;
+        $this->trustedControlEndpointCertificateName = $trustedControlEndpointCertificateName;
+        $this->trustedCacheEndpointCertificateName = $trustedCacheEndpointCertificateName;
+    }
+
+    public function getAuthToken(): string
+    {
+        return $this->authToken;
+    }
+
+    public function getCacheEndpoint(): string
+    {
+        return $this->cacheEndpoint;
+    }
+
+    public function getControlEndpoint(): string
+    {
+        return $this->controlEndpoint;
+    }
+
+    public function getTrustedControlEndpointCertificateName(): string|null
+    {
+        return $this->trustedControlEndpointCertificateName;
+    }
+
+    public function getTrustedCacheEndpointCertificateName(): string|null
+    {
+        return $this->trustedCacheEndpointCertificateName;
     }
 
     /**
